cert: Fix verification of the 2nd certificate in chain
In the case that the trusted CA certificates are provided to
l_certchain_verify() and one of them binary-matches the top CA
certificate in the chain being verified, we'd skip linking it to the
keyring and we'd link the next certificate directly.  As a result
that second certificate would be linked to an empty unrestricted keyring
and would not be verified at all.

Instead add the top CA to the keyring and restrict it the same way we
proceed when no trusted CAs are provided.