cert: Relax trused CA matching in l_certchain_verify
Check if any of the certificates in the chain matched any of the
trusted CA certificates (if given) and treat the chain as trusted even
if the matching certificate was not the top CA in the chain.
We also don't require that the trusted certificates or the root in the
chain are self-signed.

This allows a user to pin a specific certificate to a server/client or
trust a subtree of all of the certificates emitted by the root CA.