dhcp6: change option code to uint16_t
GCC-10 appears to have gotten "smarter" (or pickier) and passing
unnamed enums as an argument expecting a named enum results in
an error. The internal APIs in dhcp6 have been changed to take a
uint16_t instead of enum l_dhcp6_option. This allows both public
and internal DHCP options to be passed to these builder APIs.