io: Call disconnect handler before removing the watch
When encountering EPOLLERR or EPOLLHUP conditions, first invoke
disconnect handler if it is set, and only then clear/remove the watch.

This allows the client to retrieve the actual code for the socket
error via getsockopt(). Otherwise, if the socket is closed prior
to calling the disconnect handler, the error code is lost.