acd: add probe/annouce states, and announce address
Tracks state (probe/annouce) and will stop probing if an address
conflict is detected. Once all probes have been sent the state
transitions to announce. At this point the IP is safe to use.

Events are now signaled for conflicts and available IPs.

Going forward, the ANNOUNCED state will be the normal state the
ACD client is in while monitoring for conflicting packets.