tls: Refactor compression method ID check
When validating the Client Hello message we look up the compression
method IDs passed by the client in our look-up table.  After this is
done check that at least one ID was found in the table, rather than
use memchr() before that lookup to make sure it contains the only ID
that we support.  This should be less confusing to static analysis.