dhcp-server: Add API for manipulating leases
Add methods for obtaining and releasing DHCP leases.  There are four
methods that correspond to the DISCOVER, REQUEST, DECLINE and RELEASE
messages and two shortcut methods for releasing the leases.  This API is
useful for implementing mechanisms like the optional P2P IP Allocation
in the 802.11 4-Way Handshake and FILS IP Address Allocation in 802.11
Associate frames (both reduces the number of roundtrips in connection
setup).  Since there are multiple such mechanisms I decided to include
discover and request as separate methods to allow more flexibility.