dhcp-server: Save lease mac before calling add_lease
add_lease will re-use the existing lease struct but will memzero it and
the fill in the data.  We can't pass lease->mac to add_lease() because
the mac will be zeroed together with the rest of the structure and will
be all-zeros when add_lease() tries to copy the new value from the
pointer we passed.